This commit was generated by cvs2svn to compensate for changes in r5977, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@5978 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/unix/x11/xaw/uimenu.c
@@ -468,7 +468,8 @@ void _ui_menu_radio_helper(Widget w,
                                 (resource_value_t) client_data);
             ui_update_menus();
 	    if (psid_mode) {
-	        sprintf(buf, "resource %s %d\n", resource_name, client_data);
+	        sprintf(buf, "resource %s %d\n", resource_name,
+                        (int)client_data);
 		ui_proc_write_msg(buf);
 	    }
         }
